refactor: relocate core logic to pkg subpackages
- Move domain models (Stock, Ticker) to pkg/models.
- Move Naver API client and scraper logic to pkg/naver.
- Move ticker search functionality to pkg/search.
- Move storage repositories and CSV data to pkg/storage.
- Move path configuration to pkg/config.
- Update all internal package imports and references.
