+++ internal/cli/market.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/ericyhkim/juga/internal/core"
 	"github.com/ericyhkim/juga/internal/ui"
+	"github.com/ericyhkim/juga/pkg/naver"
 
 	"github.com/spf13/cobra"
 )
@@ -16,7 +16,7 @@ var marketCmd = &cobra.Command{
 	Short:   "Show detailed market index information",
 	Long:    `Display detailed statistics for KOSPI and KOSDAQ, including high/low prices and trading volume/value.`,
 	Run: func(cmd *cobra.Command, args []string) {
-		client := core.NewClient()
+		client := naver.NewClient()
 		indices, err := client.FetchIndices()
 		if err != nil {
 			fmt.Fprintf(os.Stderr, "Error fetching market data: %v\n", err)
