+++ pkg/config/paths.go
