+++ pkg/models/ticker.go
@@ -1,0 +1,7 @@
+package models
+
+type Ticker struct {
+	Code   string
+	Name   string
+	Market string
+}
