+++ pkg/naver/encoding_test.go
@@ -1,4 +1,4 @@
-package core
+package naver
 
 import (
 	"testing"
